refactor: update paths in lambda functions
